requirements.txt
numpy==2.4.2
pandas==3.0.2
biopython==1.86
torch==2.8.0; python_version < "3.14"
torch==2.9.0; python_version >= "3.14"
streamlit==1.56.0
